1101-1

+++ forum/pages/api/post/del.js
@@ -1,19 +1,25 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(req, res) {
     if (req.method == "POST") {
-        try {
-            let db = (await connectDB).db("forum");
+        console.log(req.body);
+        let session = await getServerSession(req, res, authOptions);
+
+        const db = (await connectDB).db("forum");
+        let 찾은거 = await db
+            .collection("post")
+            .findOne({ _id: new ObjectId(req.body) });
+
+        if (찾은거.author == session.user.email) {
             let result = await db
                 .collection("post")
                 .deleteOne({ _id: new ObjectId(req.body) });
-
-            // 성공적인 응답을 JSON 형식으로 반환
-            res.status(200).json({ message: "Success", result });
-        } catch (error) {
-            // 오류 응답을 JSON 형식으로 반환
-            res.status(500).json({ error: "Internal Server Error" });
+            return res.status(200).json("삭제완료");
+        } else {
+            return res.status(500).json("현재유저와 작성자 불일치");
         }
     }
 }
